stupid dumb fixes

+++ Assets/Scripts/Enemy/EnemyShieldAttack.cs
@@ -5,7 +5,7 @@ public class EnemyShieldAttack : MonoBehaviour
 {
 
     private Transform body;
-    private int dmg;
+    private float dmg;
     private Collider myCol;
     void Start()
     {
